Move modules js files to modules folder
