+++ src/modules/check.js
